Initial commit from dk
This includes a number of things:
-simpleframework breaks down each robot behavior into a separate class.
The core method in each class is doTheThing(), which does things. Right
now, it makes a bunch of beavers, miners, and a few soldiers. It
distributes supply, and some of the units attack.
-betterframework also breaks down each behavior into a subclass. I think
it's more extensible, but comes at the cost of several hundred
bytecodes. right now it just makes beavers, but it also does distributed
BFS pathfinding with spare bytecodes (in progress).

also neither of these work work release 1.02, because java collections
are borked in that release.