Added launcher actions in dronerush
Also moved the revamped pathfinding over to dronerush. launchers >
towers.

Right now the launchers and missiles micro effectively, but their
pathfinding is still based on distance-from-our-hq. We could improve
this by 1. inferring wether the map is a rotation or a translation and
then 2. rotating or reflecting the pathfinding results to get the
distance from the *enemy's* hq.