Added pathing-to-enemy-hq to betterframework
unfortunately, this also introduced a bunch of bugs. ironing those out
in the next few commits.

the pathing is okay, but it would benefit very directly from having a
better search implementation. since apparently robots are run
sequentially, a distributed BFS should be able to work, if we just
re-implement it.