Added some symmetry-based logic in pathfinding
Now launchers shouldn't get stuck in corners of the map, as we do
pathfinding both on squares we can see AND on squares we can't see, but
can infer based on map reflections.

This doesn't seem to be a burden on the pathfinding queue--in fact, it
might do just the opposite, since it allows us to explore "UNKNOWN"
tiles without enqueueing those tiles to check again later.