Rough fix for bug mode cycles.
I don't think a complete fix is feasible given bytecode restrictions; a
complete fix would be to store a set of past positions, and the actions
we took at those positions, and we just don't have the bytecodes for
that. instead, this just stores the past 10 positions, which sort of
helps a little.